Fixed a bug where worlds wouldn't get unlocked on unload. Partially implemented 1.14 support.
So, I felt that I had to try again, and I decided to took the javassist approach. Although it's a bit messy (and I don't like the idea of having to edit the startup parameters of the server just for this), I've managed to override the chunk load method. I'll deal with the saving method later. Also had to update the slime format to v5, as skylight and blocklight are not always present anymore.

+++ slimeworldmanager-plugin/src/main/java/com/grinderwolf/smw/plugin/loaders/FileLoader.java
@@ -56,7 +56,7 @@ public class FileLoader implements SlimeLoader {
     }
 
     @Override
-    public void saveWorld(String worldName, byte[] serializedWorld) throws IOException {
+    public void saveWorld(String worldName, byte[] serializedWorld, boolean lock) throws IOException {
         File lastBackup = new File(WORLD_DIR, worldName + ".slime_old");
 
         if (lastBackup.exists()) {
@@ -75,23 +75,28 @@ public class FileLoader implements SlimeLoader {
             lastBackup.delete();
         }
 
-        // Make sure the lock file is there
-        File lockFile = new File(WORLD_DIR, worldName + ".slime_lock");
+        if (lock) {
+            // Make sure the lock file is there
+            File lockFile = new File(WORLD_DIR, worldName + ".slime_lock");
 
-        if (!lockFile.exists()) {
-            lockFile.createNewFile();
+            if (!lockFile.exists()) {
+                lockFile.createNewFile();
 
-            try (FileOutputStream fileStream = new FileOutputStream(lockFile);
-                 DataOutputStream dataStream = new DataOutputStream(fileStream)) {
-                dataStream.writeLong(System.currentTimeMillis());
+                try (FileOutputStream fileStream = new FileOutputStream(lockFile);
+                     DataOutputStream dataStream = new DataOutputStream(fileStream)) {
+                    dataStream.writeLong(System.currentTimeMillis());
+                }
             }
         }
     }
 
     @Override
-    public void unlockWorld(String worldName) {
-        File lockFile = new File(WORLD_DIR, worldName + ".slime_lock");
+    public void unlockWorld(String worldName) throws UnknownWorldException {
+        if (!worldExists(worldName)) {
+            throw new UnknownWorldException(worldName);
+        }
 
+        File lockFile = new File(WORLD_DIR, worldName + ".slime_lock");
         lockFile.delete();
     }
